+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Unreleased
   in parent
 - Sealed `Cap` and `MutCap` traits
 - Added a step building the documentation to CI pipeline
-- Bumped minimum supported Rust version to `1.46`
+- Bumped minimum supported Rust version to `1.49`
 
 
 0.6.0