+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Sealed `Cap` and `MutCap` traits
 - Added a step building the documentation to CI pipeline
 - Switched to using GitHub Actions as CI provider
+  - Added GitHub Actions workflow for publishing the crate
 - Bumped minimum supported Rust version to `1.58`